% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_and_predict.R
\name{data_for_prediction}
\alias{data_for_prediction}
\title{Model and Visualize Sneaker Prices}
\usage{
data_for_prediction(table1, size)
}
\arguments{
\item{table1}{Table containing sneaker pricing information}

\item{size}{Size of the Sneaker}
}
\description{
Given a sneaker table and sneaker size, this routine attempts to 
create a model for future prices and visualize it simultaneously.
}
\details{
This function is for making predictions and plotting the predictions
given by the table after calculating mean value of shoes in the same day and
same size. Result will consists of three parts, first one is plot
second and third one are strings indicating the result of maximum value and
minimum value separately
}
