% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_and_predict.R
\name{data_for_visualization}
\alias{data_for_visualization}
\title{Visualize a Single Sneaker Prices}
\usage{
data_for_visualization(table)
}
\arguments{
\item{table}{Pricing information}
}
\value{
A `list` with a `ggplot2` object.
}
\description{
This routine visualizes only one type of sneaker from its original data.
}
