% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_and_predict.R
\name{data_for_visualization2}
\alias{data_for_visualization2}
\title{Visualize Two Sneaker Prices}
\usage{
data_for_visualization2(name1, name2)
}
\arguments{
\item{name1, name2}{Names of the sneakers to use for visualization.}
}
\value{
A `list` with a `ggplot2` object.
}
\description{
This visualizes compares sales of two kind of shoes
}
