% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_info}
\alias{get_info}
\title{Retrieves Sneaker Information}
\usage{
get_info(remDr)
}
\arguments{
\item{remDr}{An `RSelenium` remote driver}
}
\description{
Downloads the desired sneaker prices and picture from Website.
}
