% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_picture}
\alias{get_picture}
\title{Retrieve Sneaker Picture}
\usage{
get_picture(remDr)
}
\arguments{
\item{remDr}{An `RSelenium` remote driver}
}
\description{
Downloads the desired sneaker picture from Website.
}
