% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_table}
\alias{get_table}
\title{Retrieve Sneaker Price Table}
\usage{
get_table(remDr)
}
\arguments{
\item{remDr}{An `RSelenium` remote driver}
}
\value{
A `data.frame` with the sneaker pricing information
}
\description{
Obtains Sneaker Pricing Information
}
