% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_rawdata.R
\name{handle_rawdata}
\alias{handle_rawdata}
\title{Import Raw Data in SQL Data Base}
\usage{
handle_rawdata(input)
}
\arguments{
\item{input}{The name of the file without the `.csv` extension.}
}
\value{
A `data.frame` and have the side effect of writing a new table to
disk.
}
\description{
This function is used for handle raw data, add variables, and extract value by SQL to make predictions.
}
