% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sneakers.R
\name{run_sneakers}
\alias{run_sneakers}
\title{Launch Sneakers Shiny Application}
\usage{
run_sneakers()
}
\description{
Opens an interactive Shiny application that shows current sneaker prices
and attempts to forecast future prices.
}
