% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_sneaker_info}
\alias{scrape_sneaker_info}
\title{Scrape Sneaker Pricing}
\usage{
scrape_sneaker_info(name)
}
\arguments{
\item{name}{The name of a sneaker.}
}
\value{
A `list` with pricing and picture information.
}
\description{
Retrieves sneaker pricing from stockx.com
}
