% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{search_website}
\alias{search_website}
\title{Search the Sneaker Website for Sneaker Data}
\usage{
search_website(web, shoe, remDr)
}
\arguments{
\item{web}{The website address}

\item{shoe}{The name of the shoe}

\item{remDr}{An `RSelenium` remote driver}
}
\description{
Triggers a search on the website for sneaker data
}
