% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sneakers-package.R
\docType{package}
\name{sneakers-package}
\alias{sneakers}
\alias{sneakers-package}
\title{sneakers: Obtain and Forecast Sneaker Market Data}
\description{
Using either a pre-built cache or retrieving data in real-time,
    the package allows for sneaker aficionados to have an overview of the current
    sneaker market.
}
\details{
Group project for STAT 385 that attempts to retrieve sneaker price data.
}
\author{
\strong{Maintainer}: Ziwei Liu \email{ziweil2@illinois.edu}

Authors:
\itemize{
  \item Zepeng Xiao \email{zepengx2@illinois.edu}
  \item Yuquan Zheng \email{yzheng58@illinois.edu}
}

Other contributors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu} [contributor]
}

}
